Potential fix for code scanning alert no. 46: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (size_t) to avoid overflow. This can be done by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using size_t, which has a larger range and can prevent overflow.

The specific change needed is to cast mb_w to size_t before multiplying it by mb_h on line 156.

+++ libwebp-1.0.3/src/enc/webp_enc.c
@@ -153,7 +153,7 @@ static VP8Encoder* InitVP8Encoder(const WebPConfig* const config,
   const size_t preds_size = (size_t)preds_w * preds_h * sizeof(*enc->preds_);
   const int top_stride = mb_w * 16;
   const size_t nz_size = (mb_w + 1) * sizeof(*enc->nz_) + WEBP_ALIGN_CST;
-  const size_t info_size = mb_w * mb_h * sizeof(*enc->mb_info_);
+  const size_t info_size = (size_t)mb_w * mb_h * sizeof(*enc->mb_info_);
   const size_t samples_size =
       2 * top_stride * sizeof(*enc->y_top_)  // top-luma/u/v
       + WEBP_ALIGN_CST;                      // align all
